requirements.txt
jsonpickle~=1.4, >= 1.4.1
requests~=2.25
cachecontrol~=0.12.6
python-dateutil~=2.8.1
enum34~=1.1, >=1.1.10
deprecation~=2.1
